Data.BinaryList.Algorithm.BinaryTransform

Description

Implementation of the binary transform, as detailed in https://github.com/plow-technologies/writings/tree/master/binary-transform.

Bijections

data Bijection a b

A bijection from a to b is a function from a to b that is invertible. A function is invertible if and only if it's both injective and surjective. These are the definitions of the terms injective and surjective.

A function f :: a -> b is injective if f(x) = f(y) implies x = y for every x, y :: a.

A function f :: a -> b is surjective if for every y :: b there is x :: a such that f(x) = y.

To apply a bijection f to an argument x use direct f x. To apply its inverse just do inverse f x.

Constructors

Bijection

Fields

direct :: a -> b

Bijection.

inverse :: b -> a

Inverse of the bijection.

inverseBijection :: Bijection a b -> Bijection b a

The inverse of a bijection.

functorBijection :: Functor f => Bijection a b -> Bijection (f a) (f b)

Lift a Bijection to work over an arbitrary Functor.

productBijection :: Bijection a b -> Bijection c d -> Bijection (a, c) (b, d)

The product of two bijections. This is the equivalent to *** for the Bijection type.

Binary Transform

Left version

leftBinaryTransform :: Bijection (a, a) (a, a) -> Bijection (BinList a) (BinList a)

The left binary transform lifts a permutation (i.e. a bijection from a set to itself) of a plane to a permutation of binary lists. The transformation condenses at the left.

leftInverseBinaryTransformDec :: Bijection (a, a) (a, a) -> Decoded a -> Decoded a

leftPartialInverse :: Bijection (BinList a) (BinList a) -> Int -> BinList a -> BinList a

Apply the inverse of a permutation of binary lists to a sublist of a binary list. The Int argument specifies the size of the sublist. More specifically, applying leftPartialInverse f i to a binary list xs of length 2^n returns the result of applying inverse f to the first max{1,2^(n-i)} elements.

Right version

rightBinaryTransform :: Bijection (a, a) (a, a) -> Bijection (BinList a) (BinList a)

The right binary transform lifts a permutation (i.e. a bijection from a set to itself) of a plane to a permutation of binary lists. The transformation condenses at the right.

rightInverseBinaryTransformDec :: Bijection (a, a) (a, a) -> Decoded a -> Decoded a

rightPartialInverse :: Bijection (BinList a) (BinList a) -> Int -> BinList a -> BinList a

Apply the inverse of a permutation of binary lists to a sublist of a binary list. The Int argument specifies the size of the sublist. More specifically, applying rightPartialInverse f i to a binary list xs of length 2^n returns the result of applying inverse f to the last max{1,2^(n-i)} elements.